rename Viewer to TestViewer
